Prep for search results view
- Make input own component
- Adjust navigation styles
- Add clear input functionality
- Mobile style improvements
